Need to subtract signed values otherwise the value wraps instead of going below zero.  Was causing the stream to hang when using the resampler.

+++ xbmc/cores/AudioEngine/Engines/SoftAEStream.cpp
@@ -443,7 +443,7 @@ unsigned int CSoftAEStream::ProcessFrameBuffer()
   }
 
   if (m_refillBuffer)
-    m_refillBuffer = std::max((unsigned int)m_refillBuffer - frames, (unsigned int)0);
+    m_refillBuffer = (unsigned int)std::max((int)m_refillBuffer - (int)frames, 0);
 
   return consumed;
 }
